test: split cli domain coverage and tighten docs
